Replace a single TestDoubles trait with mock framework specific ones
