+++ tests/TestCase/TestDoubles/NotNull/ProphecyTest.php
@@ -6,7 +6,7 @@ namespace Zalas\PHPUnit\Doubles\Tests\TestCase\TestDoubles\NotNull;
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Prophecy\Prophecy\ObjectProphecy;
-use Zalas\PHPUnit\Doubles\TestCase\TestDoubles;
+use Zalas\PHPUnit\Doubles\TestCase\ProphecyTestDoubles;
 use Zalas\PHPUnit\Doubles\Tests\TestCase\TestDoubles\Fixtures\Copper;
 use Zalas\PHPUnit\Doubles\Tests\TestCase\TestDoubles\Fixtures\Fred;
 use Zalas\PHPUnit\Doubles\Tests\TestCase\TestDoubles\Fixtures\Nobby;
@@ -14,7 +14,7 @@ use Zalas\PHPUnit\Doubles\Tests\TestCase\TestDoubles\Fixtures\Nobby;
 class ProphecyTest extends TestCase
 {
     use ProphecyTrait;
-    use TestDoubles;
+    use ProphecyTestDoubles;
 
     /**
      * @var Nobby|ObjectProphecy
