+++ tests/TestCase/TestDoubles/ProphecyTest.php
@@ -6,7 +6,7 @@ namespace Zalas\PHPUnit\Doubles\Tests\TestCase\TestDoubles;
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Prophecy\Prophecy\ObjectProphecy;
-use Zalas\PHPUnit\Doubles\TestCase\TestDoubles;
+use Zalas\PHPUnit\Doubles\TestCase\ProphecyTestDoubles;
 use Zalas\PHPUnit\Doubles\Tests\TestCase\TestDoubles\Fixtures\Copper;
 use Zalas\PHPUnit\Doubles\Tests\TestCase\TestDoubles\Fixtures\Death;
 use Zalas\PHPUnit\Doubles\Tests\TestCase\TestDoubles\Fixtures\Discworld;
@@ -18,7 +18,7 @@ use Zalas\PHPUnit\Doubles\Tests\TestCase\TestDoubles\Fixtures\Vimes;
 class ProphecyTest extends TestCase
 {
     use ProphecyTrait;
-    use TestDoubles;
+    use ProphecyTestDoubles;
 
     /**
      * @var Vimes|ObjectProphecy
